Move VirtualNode#constructor
